renderflooding change

+++ flood/flood/grid.cpp
@@ -15,6 +15,7 @@
 #include <math.h>
 #include <time.h>
 
+//#define VUL_MARGIN 0.1
 
 
 const int NEW_WATER = -8000;
@@ -69,6 +70,10 @@ void setNotVisited(Grid* elevgrid, Grid* floodedgrid, float rise) {
             if (floodedgrid->data[i][j] == HAVENT_VISITED) {
                 floodedgrid->data[i][j] = elevgrid->data[i][j] - rise;
             }
+//            if (floodedgrid->data[i][j] <= rise + VUL_MARGIN && floodedgrid->data[i][j] >= rise - VUL_MARGIN) {
+//                floodedgrid->data[i][j] = 0;
+//            }
+             
         }
     }
 }
@@ -86,13 +91,27 @@ void setinterp_bfeNotVisited(Grid* elevgrid, Grid* interp_bfegrid,Grid* floodedg
 
                     } else {
                         floodedgrid->data[i][j] = elevgrid->data[i][j] - rise;
-
                     }
-                    
                 } else {
                     floodedgrid->data[i][j] = elevgrid->data[i][j] - rise;
                 }
             }
+            
+//            if (interp_bfegrid->data[i][j] != interp_bfegrid->NODATA_value) {
+//                if(elevgrid->data[i][j]<=(rise+ interp_bfegrid->data[i][j]) + VUL_MARGIN && elevgrid->data[i][j] >= (rise+ interp_bfegrid->data[i][j]) - VUL_MARGIN) {
+//                    floodedgrid->data[i][j] = 0;
+//                }
+//            } else {
+//                if(elevgrid->data[i][j]<=rise + VUL_MARGIN && elevgrid->data[i][j] >= rise - VUL_MARGIN) {
+//                    floodedgrid->data[i][j] = 0;
+//                }
+//            }
+
+            
+
+//            if (floodedgrid->data[i][j] <= rise + VUL_MARGIN && floodedgrid->data[i][j] >= rise - VUL_MARGIN) {
+//                floodedgrid->data[i][j] = 0;
+//            }
         }
     }
 }
@@ -152,11 +171,11 @@ void freeGridData(Grid* g) {
 
 void mallocGrid(Grid eg, Grid* vg) {
     
-    vg->data = (float**)malloc(eg.nrows * sizeof(float *));
+    vg->data = (double**)malloc(eg.nrows * sizeof(double *));
     assert(vg->data);
     
     for(int a = 0; a < eg.nrows; a++) {
-        vg->data[a] = (float*)malloc(eg.ncols * sizeof(float));
+        vg->data[a] = (double*)malloc(eg.ncols * sizeof(double));
         assert(vg->data[a]);
     }
 }
@@ -209,16 +228,16 @@ void readGridfromFile(const char* gridfname, Grid* g, double converter) {
     }
 
     
-    g->data = (float**)malloc(g->nrows * sizeof(float *));
+    g->data = (double**)malloc(g->nrows * sizeof(double *));
     assert(g->data);
     for(int i = 0; i < g->nrows; i++) {
-        g->data[i] = (float*)malloc(g->ncols * sizeof(float));
+        g->data[i] = (double*)malloc(g->ncols * sizeof(double));
         assert(g->data[i]);
     }
     
     for(int i = 0; i < g->nrows; i++) {
         for(int j = 0; j < g->ncols; j++) {
-            int y = fscanf(f,"%f ",&g->data[i][j]);
+            int y = fscanf(f,"%lf ",&g->data[i][j]);
             if (y != 1) {
                 printf("ERROR: fscanf did not properly scan in the grid\n");
                 exit(1);
